requirements.txt
google-generativeai
ddgs
requests==2.32.5
beautifulsoup4==4.14.3
python-dotenv
